Migrate tests from Mocha+Chai to Jest
Using Jest instead of the previous setup will allow for the tests to have further maintainability in the future, as Jest is more widely used.
Furthermore, Jest is also a more complete solution, providing coverage and mocking without having to rely on external packages.
The fact that it provides built-in mocking will allow for the creation of better unit tests, without having to use only end-to-end tests as was being done until now.

Closes #25.