Add mock data management script
The mock data and the management script both live inside folder mock.
The mock data is composed of two files:

	offers.json
	companies.json

The scraper formats this data as an array for easier insertion into
mongodb and also establishes connections between offers and companies.
The script needs to handle the ownership connection after insertion
into mongodb, i.e. needs to assign the owner company to each offer.

The script is to be called at the root of the repository, with an
optional command. For example (one of):

	node mock clean
	node mock populate
	node mock

where the last command both cleans and populates.